requirements.txt
appnope==0.1.2
autopep8==1.5.7
backcall==0.2.0
debugpy==1.4.1
decorator==5.0.9
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytube==10.9.3
pyzmq==22.1.0
six==1.16.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
